refactor: logProvider contract was moved to shared application folder
